+++ uc/hal_fake/hal_fake.c
@@ -61,7 +61,7 @@ TF_HalCommon *tf_hal_get_common(TF_HalContext *hal) {
     return &hal->hal_common;
 }
 
-void tf_hal_log_message(const char *msg, uint32_t len) {
+void tf_hal_log_message(const char *msg, size_t len) {
     (void) msg;
     (void) len;
 }
